Post a review when configured to do so. (#5)
## Before this PR
The only way to run this check is to fail the check when requirements have not been satisfied.

## After this PR
Add an action option 'post-review' that when set to 'true' will cause this task to post a review to the PR approving when requirements are satisfied or requesting changes when requirements have not been satisfied. When requirements have not been satisfied and a request changes review is posted, the task will still mark itself as succeeded.

Compared to the prior behavior, this will make the PR summary view a little easier to digest.

## Possible downsides?
<!-- Please describe any way users could be negatively affected by this PR. -->